+++ packages/blockchain-link/src/workers/stellar/types.ts
@@ -1,0 +1,10 @@
+import type { TokenDetailByMint } from '@trezor/blockchain-link-types';
+import type { StellarAPI } from '@trezor/network-stellar/types';
+
+import type { ContextType } from '../baseWorker';
+
+export type Context = ContextType<StellarAPI> & {
+    getTokenMetadata: () => Promise<TokenDetailByMint>;
+};
+
+export type Request<T> = T & Context;
